split code in imports
